+++ apps/frontend/src/libs/modules/store/libs/types/extra-arguments.type.ts
@@ -1,6 +1,7 @@
 import { type Storage } from "~/libs/modules/storage/storage.js";
 import { type ToastNotifier } from "~/libs/modules/toast-notifier/toast-notifier.js";
 import { type authApi } from "~/modules/auth/auth.js";
+import { type contributorApi } from "~/modules/contributors/contributors.js";
 import { type groupApi } from "~/modules/groups/groups.js";
 import { type permissionApi } from "~/modules/permissions/permissions.js";
 import { type projectApiKeysApi } from "~/modules/project-api-keys/project-api-keys.js";
@@ -11,6 +12,7 @@ import { type userApi } from "~/modules/users/users.js";
 
 type ExtraArguments = {
 	authApi: typeof authApi;
+	contributorApi: typeof contributorApi;
 	groupApi: typeof groupApi;
 	permissionApi: typeof permissionApi;
 	projectApi: typeof projectApi;