+++ apps/frontend/src/modules/project-groups/slices/actions.ts
@@ -5,7 +5,10 @@ import {
 	type AsyncThunkConfig,
 	type PaginationQueryParameters,
 } from "~/libs/types/types.js";
-import { type UserGetAllResponseDto } from "~/modules/users/users.js";
+import {
+	type UserGetAllQueryParameters,
+	type UserGetAllResponseDto,
+} from "~/modules/users/users.js";
 
 import {
 	type ProjectGroupCreateRequestDto,
@@ -16,7 +19,7 @@ import { name as sliceName } from "./project-group.slice.js";
 
 const loadUsers = createAsyncThunk<
 	UserGetAllResponseDto,
-	PaginationQueryParameters,
+	UserGetAllQueryParameters,
 	AsyncThunkConfig
 >(`${sliceName}/load-users`, (query, { extra }) => {
 	const { userApi } = extra;
@@ -34,6 +37,23 @@ const loadAllByProjectId = createAsyncThunk<
 	return await projectGroupApi.getAllByProjectId(projectId, query);
 });
 
+const deleteById = createAsyncThunk<boolean, { id: number }, AsyncThunkConfig>(
+	`${sliceName}/delete-by-id`,
+	async ({ id }, { extra }) => {
+		const { projectGroupApi, toastNotifier } = extra;
+
+		const isDeleted = await projectGroupApi.deleteById(id);
+
+		if (isDeleted) {
+			toastNotifier.showSuccess(
+				NotificationMessage.PROJECT_GROUP_DELETE_SUCCESS,
+			);
+		}
+
+		return isDeleted;
+	},
+);
+
 const create = createAsyncThunk<
 	ProjectGroupGetAllItemResponseDto,
 	ProjectGroupCreateRequestDto,
@@ -48,4 +68,4 @@ const create = createAsyncThunk<
 	return response;
 });
 
-export { create, loadAllByProjectId, loadUsers };
+export { create, deleteById, loadAllByProjectId, loadUsers };