+++ packages/shared/src/index.ts
@@ -59,6 +59,7 @@ export {
 	ContributorError,
 	type ContributorGetAllItemResponseDto,
 	type ContributorGetAllResponseDto,
+	ContributorsApiPath,
 } from "./modules/contributors/contributors.js";
 export {
 	type GitEmailCreateRequestDto,